move original index.js to client. add root index.js for server start
