+++ index.js
@@ -1,48 +1,7 @@
-import { createStore } from 'redux'
 
-const initialWombatState = {
-  wombats: ['Gertrude', 'Bartholemew']
-}
+var app = require('./server')
+var PORT = process.env.PORT || 3000
 
-const wombatReducer = (state = initialWombatState, action) => {
-  switch (action.type) {
-    case 'ADD_WOMBAT':
-      return {
-        wombats: [...state.wombats, action.wombat]
-      }
-    case 'DEL_WOMBAT':
-      return {
-        wombats: state.wombats.filter((wombat) => wombat !== action.wombat)
-      }
-    default:
-      return state
-  }
-}
-
-const store = createStore(wombatReducer,
-  window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
-)
-
-document.addEventListener('DOMContentLoaded', () => {
-  render()
-  store.subscribe(render)
+app.listen(PORT, function () {
+  console.log('server listening on port: ', PORT)
 })
-
-function render () {
-  const state = store.getState()
-  const wombats = state.wombats
-  document.getElementById('app').innerHTML = renderWombats('wombats', wombats)
-}
-
-function renderWombats (name, wombats) {
-  let output = `<ul>`
-  for (const wombat of wombats) {
-    output += `
-    <li>
-      ${wombat}
-    </li>`
-  }
-  output += `</ul>`
-  return output
-}
-
